Move the new style example to the main package
