+++ src/test/java/examples/simple/SimpleTableMapperNewStyleTest.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package examples.simple.newstyle;
+package examples.simple;
 
 import static examples.simple.legacy.SimpleTableDynamicSqlSupport.*;
 import static org.assertj.core.api.Assertions.assertThat;
